Removed old version of distribution contract
