+++ vtxdistribut/resources/vtxdistribut-paycore-rc.md
